change 01

+++ Program.cs
@@ -6,21 +6,21 @@ int Prompt(string message)
     Console.Write(message);
     return int.Parse(Console.ReadLine());
 }
-int [,] GenerateArray(int row, int column, int min, int max)
+double [,] GenerateArray(int row, int column, int min, int max)
 {
-    var array = new int [row, column];
+    var array = new double [row, column];
     var rnd = new Random();
     for (int i = 0; i < array.GetLength(0); i++)
     {
         for (int j = 0; j < array.GetLength(1); j++)
         {
-            array[i, j] = rnd.Next(min, max + 1);
+            array[i, j] = rnd.Next (min, max + 1);
         }
     }
     return array;
 }
 
-void PrintArray(int [,] array)
+void PrintArray(double [,] array)
 {
     for (int i = 0; i < array.GetLength(0); i++)
     {
@@ -32,7 +32,7 @@ void PrintArray(int [,] array)
     }
 }
 
-double [] FindArithmeticmeanOfColumn (int [,] array)
+double [] FindArithmeticmeanOfColumn (double [,] array)
 {
     double [] AvarageArray = new double [array.GetLength(1)];
     for (int j = 0; j < array.GetLength(1); j++)
@@ -48,7 +48,7 @@ int row = Prompt("Введите количество строк: ");
 int column = Prompt("Введите количество столбцов: ");
 int min = 1;
 int max = 10;
-int [,] myarray = GenerateArray(row, column, min, max);
+double [,] myarray = GenerateArray(row, column, min, max);
 PrintArray(myarray);
 double[] AvarageArray = FindArithmeticmeanOfColumn(myarray);
 System.Console.WriteLine($"Среднее арифмитическое каждого столбца:");
